requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
pydantic==2.9.2
pypdf==4.3.1
llama-cpp-python==0.3.5
python-multipart==0.0.12
